requirements.txt
Flask==2.3.2
pydantic==1.10.2
joblib==1.2.0
pandas==1.5.3
flask-cors==3.0.10
pymongo==4.3.3
python-dotenv==1.0.0